xserver pixman fontconfig mesa plink xorgproto git update Apr 5 2020
xserver          commit 5684d436e2b65cd0fe305460e437a2f69af29865
libxcb           commit be1745c8eb00defcb31d336ccc142de056e92bd8
libxcb/xcb-proto commit dbf346502a31c3629b2c597711fb4b0ea6521de6
xkeyboard-config commit 77e49674e2344e2c6ffc4b76ec4d33f97a40a73e
libX11           commit b8766a438461e20b5c6f7a6111b703d4cb71501f
libXdmcp         commit bc141f304712a109b6b518724e1feb66293bf472
libXext          commit ebb167f34a3514783966775fb12573c4ed209625
libfontenc       commit 2baea13978759d1a011fc6d739465893b554d30a
libXinerama      commit c5187f076d16601c15c59c5a2f05c0513d9f042b
libXau           commit d9443b2c57b512cfb250b35707378654d86c7dea
xkbcomp          commit 7ffec6583ae92fe81402870eb89dea21ca34e933
pixman           commit 348e99b52f41caf32f8b7840e53f0b1322716c47
mkfontscale      commit 79f686765238ddd0e4753f9d4ee62b5a5296610e
xwininfo         commit d19200c0b0ff638edbb16555e711d2353625bd4a
libXft           commit a266847d3c17dcdfcac719a1aa970ad54f4b545a
libXmu           commit e9efe2d027b4c46cf6834cc532222f8ad1d1d3c3
libxtrans        commit 3b5df889f58a99980a35a7b4a18eb4e7d2abeac4
fontconfig       commit 9c46ef4aac87c42d013d0e7380b6aeb03e1a9949
mesa             commit 7af813d48a55a14401df51870feddfaa418397e1
putty            commit f9a46a958157e1d6bd126560623599ed8ee50f32
pthreads         commit 67186579ee3cf81e54d149dce5f5a612f957d2fe
EGL-Registry     commit 11478904448bbdf5757b798c856a525aa2b351b1
xorgproto        commit da9d1cb7a516992bf7e2e254a507848f0a38fad3

+++ mesalib/src/gallium/auxiliary/driver_rbug/rbug_core.c
@@ -44,8 +44,8 @@
 
 #include <errno.h>
 
-#define U642VOID(x) ((void *)(unsigned long)(x))
-#define VOID2U64(x) ((uint64_t)(unsigned long)(x))
+#define U642VOID(x) ((void *)(uintptr_t)(x))
+#define VOID2U64(x) ((uint64_t)(uintptr_t)(x))
 
 #define container_of(ptr, type, field) \
    (type*)((char*)ptr - offsetof(type, field))
